+++ package.json
@@ -29,7 +29,8 @@
     "radix-vue": "^1.9.2",
     "tailwind-merge": "^2.4.0",
     "tailwindcss-animate": "^1.0.7",
-    "vue": "^3.4.35"
+    "vue": "^3.4.35",
+    "vue-router": "4"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.24.0",
