+++ src/components/view/ProfileView.vue
@@ -1,0 +1,11 @@
+<!--
+ * @author: Caven
+ * @since: 2024-08-01
+-->
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div>Profile View</div>
+</template>
